feat: updated pages structure and added role based access to api routes (#31)

+++ pages/api/frontend/v0.1/orgs/[orgId]/apps/[appId]/index.ts
@@ -16,9 +16,42 @@ export default async function handler(
         return;
     }
 
+    const email = session.user?.email as string;
+
+    const user = await prisma.user.findFirst({
+        where: {
+            email: email
+        }
+    });
+
+    if (!user || ( user && !user.id)) {
+        res.status(400).json({ message: 'User not found!' });
+        return;
+    }
+    
+    const userInOrg = await prisma.usersInOrganisations.findFirst({
+        where: {
+            user: {
+                id: user.id
+            },
+            org: {
+                id: Number(req.query.orgId)
+            },
+        },
+        select: {
+            role: true
+        }
+    });
+
+    if (userInOrg?.role !== "ADMIN" && userInOrg?.role !== "USER") {
+        // if user has no business with this organisation, return a 404
+        res.status(404).json({ message: 'no organisation found with id ' + req.query.orgId });
+        return;
+    }
+
     switch (req.method) {
         case 'GET':
-            const app = await prisma.app.findUnique({
+            const app = await prisma.app.findFirst({
                 include: {
                     messages: {
                         include: {
@@ -35,19 +68,25 @@ export default async function handler(
                     }
                 },
                 where: {
-                    id: Number(req.query.appId)
+                    id: Number(req.query.appId),
+                    orgId: Number(req.query.orgId),
                 }
             })
 
             if (app == null) {
-                res.status(404).end('no app found with id ' + req.query.appId)
+                res.status(404).json({ message: 'no app found with id ' + req.query.appId });
+                return;
             }
 
-            res.status(200).json(app)
-            break
+            res.status(200).json(app);
+            break;
 
         case 'DELETE':
             try {
+                if (userInOrg?.role === "USER") {
+                    res.status(403).json({ message: 'you are not allowed to delete app with id ' + req.query.orgId });
+                    return;
+                }
                 const deletedApp = await prisma.app.delete({
                     where: {
                         id: Number(req.query.appId)
@@ -57,13 +96,17 @@ export default async function handler(
                 res.status(200).json(deletedApp)
             } catch(e) {
                 if (e instanceof Prisma.PrismaClientKnownRequestError) {
-                    res.status(404).end('no app found with id ' + req.query.appId)
+                    res.status(404).json({ message: 'no app found with id ' + req.query.appId });
                 }
             }
-            break
+            break;
 
         case 'PUT':
             try {
+                if (userInOrg?.role === "USER") {
+                    res.status(403).json({ message: 'you are not allowed to update app with id ' + req.query.orgId });
+                    return;
+                }
                 const updatedApp = await prisma.app.update({
                     where: {
                         id: Number(req.query.appId)
@@ -71,18 +114,18 @@ export default async function handler(
                     data: {
                         name: req.body.name
                     }
-                })
+                });
 
                 res.status(201).json(updatedApp)
             } catch(e) {
                 if (e instanceof Prisma.PrismaClientKnownRequestError) {
-                    res.status(404).end('no app found with id ' + req.query.appId)
+                    res.status(404).json({ message: 'no app found with id ' + req.query.appId });
                 }
             }
-            break
+            break;
 
         default:
-            res.status(405).end('method not allowed')
-            break
+            res.status(405).json({ message: 'method not allowed' });
+            return;
     }
 }
