+++ pages/api/frontend/v0.1/orgs/[orgId]/apps/[appId]/messages/[messageId].ts
@@ -28,6 +28,39 @@ export default async function handler(
         return;
     }
 
+    const email = session.user?.email as string;
+
+    const user = await prisma.user.findFirst({
+        where: {
+            email: email
+        }
+    });
+
+    if (!user || ( user && !user.id)) {
+        res.status(400).json({ message: 'User not found!' });
+        return;
+    }
+    
+    const userInOrg = await prisma.usersInOrganisations.findFirst({
+        where: {
+            user: {
+                id: user.id
+            },
+            org: {
+                id: Number(req.query.orgId)
+            },
+        },
+        select: {
+            role: true
+        }
+    });
+
+    if (userInOrg?.role !== "ADMIN" && userInOrg?.role !== "USER") {
+        // if user has no business with this organisation, return a 404
+        res.status(404).json({ message: 'no organisation found with id ' + req.query.orgId });
+        return;
+    }
+
     switch (req.method) {
         case 'GET':
             const message = await prisma.message.findUnique({
@@ -37,10 +70,11 @@ export default async function handler(
                 where: {
                     id: Number(req.query.messageId)
                 }
-            })
+            });
 
             if (message == null) {
-                res.status(404).end('no message found with id ' + req.query.messageId)
+                res.status(404).json({ message: 'no message found with id ' + req.query.messageId });
+                return;
             }
 
             res.status(200).json(message)
@@ -48,28 +82,37 @@ export default async function handler(
 
         case 'DELETE':
             try {
+                if (userInOrg?.role === "USER") {
+                    res.status(403).json({ message: 'you are not allowed to delete message with id ' + req.query.orgId });
+                    return;
+                }
                 const deletedActions = await prisma.action.deleteMany({
                     where: {
                         messageId: Number(req.query.messageId)
                     }
-                })
+                });
 
                 const deletedMessage = await prisma.message.delete({
                     where: {
                         id: Number(req.query.messageId)
                     }
-                })
+                });
 
                 res.status(200).json(deletedMessage)
             } catch(e) {
                 if (e instanceof Prisma.PrismaClientKnownRequestError) {
-                    res.status(404).end('no message found with id ' + req.query.messageId)
+                    res.status(404).json({ message: 'no message found with id ' + req.query.messageId });
+                    return;
                 }
             }
-            break
+            break;
 
         case 'PUT':
             try {
+                if (userInOrg?.role === "USER") {
+                    res.status(403).json({ message: 'you are not allowed to update message with id ' + req.query.orgId });
+                    return;
+                }
                 const updatedMessage = await prisma.message.update({
                     where: {
                         id: Number(req.query.messageId)
@@ -82,13 +125,13 @@ export default async function handler(
                         endDate: new Date(req.body.endDate),
                         appId: req.body.appId
                     }
-                })
+                });
 
                 const deletedActions = await prisma.action.deleteMany({
                     where: {
                         messageId: Number(req.query.messageId)
                     }
-                })
+                });
 
                 if (req.body.actions.length > 0) {
                     const actions: Action[] = req.body.actions;
@@ -100,16 +143,17 @@ export default async function handler(
                     });
                 }
 
-                res.status(201).json(updatedMessage)
+                res.status(201).json(updatedMessage);
             } catch(e) {
                 if (e instanceof Prisma.PrismaClientKnownRequestError) {
-                    res.status(404).end('no message found with id ' + req.query.messageId)
+                    res.status(404).json({ message: 'no message found with id ' + req.query.messageId });
+                    return;
                 }
             }
-            break
+            break;
 
         default:
-            res.status(405).end('method not allowed')
-            break
+            res.status(405).json({ message: 'method not allowed' });
+            return;
     }
 }
